cnf ran: reorganize powermanagement helpers
This moves the powermanagement stats file to be a package under ran and adds unit tests for it. Additionally, the metric collections helpers have been moved to their own package under powermanagement.
